Add file utils and update templates
